move features out of index and split loggers
