+++ src/index.js
@@ -3,18 +3,9 @@ const {
   GatewayIntentBits,
   Partials,
   EmbedBuilder,
-  PermissionsBitField,
   Collection,
   Events,
-  ActionRowBuilder,
-  ButtonBuilder,
-  ButtonStyle,
-  ModalBuilder,
-  TextInputBuilder,
-  TextInputStyle,
-  ChannelType,
 } = require("discord.js");
-const { readdirSync, readFileSync, writeFileSync } = require("fs");
 const fs = require("fs");
 const client = new Client({
   intents: [
@@ -48,10 +39,33 @@ const client = new Client({
     Partials.User,
   ],
 });
-
+exports.client = client;
 client.commands = new Collection();
 
+// anticrash
+const process = require("node:process");
+process.on("unhandledRejection", (reason, promise) => {
+  console.log("Unhandled Rejection at:", promise, "reason:", reason);
+});
+
 require("dotenv").config();
+
+require("./anti-link.js");
+require("./modals.js");
+require("./tickets.js");
+require("./bot-guilds.js");
+
+require("./logger/automod-rules.js");
+require("./logger/bans.js");
+require("./logger/channels.js");
+require("./logger/emojis.js");
+require("./logger/invites.js");
+require("./logger/members.js");
+require("./logger/messages.js");
+require("./logger/roles.js");
+require("./logger/stickers.js");
+require("./logger/threads.js");
+require("./logger/users.js");
 
 const functions = fs
   .readdirSync("./src/functions")
@@ -61,12 +75,6 @@ const eventFiles = fs
   .filter((file) => file.endsWith(".js"));
 const commandFolders = fs.readdirSync("./src/commands");
 
-// anticrash
-const process = require("node:process");
-process.on("unhandledRejection", (reason, promise) => {
-  console.log("Unhandled Rejection at:", promise, "reason:", reason);
-});
-
 (async () => {
   for (file of functions) {
     require(`./functions/${file}`)(client);
@@ -75,500 +83,6 @@ process.on("unhandledRejection", (reason, promise) => {
   client.handleCommands(commandFolders, "./src/commands");
   client.login(process.env.TOKEN);
 })();
-
-// console logging
-/*
-const logStream = fs.createWriteStream("./logs/console.txt", { flags: "a" });
-const originalLog = console.log;
-console.log = function () {
-  logStream.write(`${new Date().toISOString()} - `);
-  logStream.write(Array.from(arguments).join(" ") + "\n");
-
-  originalLog.apply(console, arguments);
-};
-*/
-
-// Ticket system
-const ticketSchema = require("./Schemas/ticketSchema");
-
-// ticket modal create
-client.on(Events.InteractionCreate, async (interaction) => {
-  if (interaction.isButton()) return;
-  if (interaction.isChatInputCommand()) return;
-
-  const modalTicket = new ModalBuilder()
-    .setTitle("Provide us with more information")
-    .setCustomId("modalTicket");
-
-  const topicTicket = new TextInputBuilder()
-    .setCustomId("topicTicket")
-    .setRequired(true)
-    .setLabel("What is the topic of your ticket?")
-    .setPlaceholder("Server connection issue? Server crashed. Etc.")
-    .setStyle(TextInputStyle.Short);
-
-  const infoTicket = new TextInputBuilder()
-    .setCustomId("infoTicket")
-    .setRequired(true)
-    .setLabel("Specify the issue, in detail.")
-    .setPlaceholder(
-      "Error messages? What happened? What have you tried already to fix it? Etc."
-    )
-    .setStyle(TextInputStyle.Paragraph);
-
-  const additionalTicket = new TextInputBuilder()
-    .setCustomId("additionalTicket")
-    .setRequired(true)
-    .setLabel("Anything to add? Links/Clips/Screenshots/etc.")
-    .setPlaceholder("Links to Clips/Screenshots. Server IP/Name.")
-    .setStyle(TextInputStyle.Paragraph);
-
-  const firstActionRow = new ActionRowBuilder().addComponents(topicTicket);
-  const secondActionRow = new ActionRowBuilder().addComponents(infoTicket);
-  const thirdActionRow = new ActionRowBuilder().addComponents(additionalTicket);
-
-  modalTicket.addComponents(firstActionRow, secondActionRow, thirdActionRow);
-
-  let choices;
-  if (interaction.isStringSelectMenu()) {
-    choices = interaction.values;
-
-    const result = choices.join("");
-
-    ticketSchema.findOne({ Guild: interaction.guild.id }).then((data) => {
-      if (!data) return;
-      const filter = { Guild: interaction.guild.id };
-      const update = { Ticket: result };
-
-      ticketSchema
-        .updateOne(filter, update, {
-          new: true,
-        })
-        .then((value) => {
-          console.log(value);
-        });
-    });
-
-    if (!interaction.isModalSubmit()) {
-      interaction.showModal(modalTicket);
-    }
-  }
-});
-
-// ticket creation
-client.on(Events.InteractionCreate, async (interaction) => {
-  if (interaction.isModalSubmit()) {
-    if (interaction.customId === "modalTicket") {
-      ticketSchema
-        .findOne({ Guild: interaction.guild.id })
-        .then(async (data) => {
-          if (!data) return;
-          const topicInput =
-            interaction.fields.getTextInputValue("topicTicket");
-          const infoInput = interaction.fields.getTextInputValue("infoTicket");
-          const additionalInput =
-            interaction.fields.getTextInputValue("additionalTicket");
-
-          if (infoInput.length > 500 || additionalInput.length > 500) {
-            return await interaction.reply({
-              content: `You have entered too much text, please shorten it and try again.`,
-              ephemeral: true,
-            });
-          }
-
-          const posChannel = await interaction.guild.channels.cache.find(
-            (c) => c.name === `ticket-${interaction.user.id}`
-          );
-          if (posChannel)
-            return await interaction.reply({
-              content: `You already have an open ticket - ${posChannel}`,
-              ephemeral: true,
-            });
-
-          const category = data.Channel;
-
-          const embed = new EmbedBuilder()
-            .setColor("#ff00b3")
-            .setTitle(`${interaction.user.username}'s Ticket`)
-            .setDescription(
-              `Thank you for opening a ticket. Please wait while the staff reviews your information. We will respond to you shortly.`
-            )
-            .addFields([
-              {
-                name: `Topic`,
-                value: `${topicInput}`,
-                inline: false,
-              },
-              {
-                name: `Info`,
-                value: `${infoInput}`,
-                inline: false,
-              },
-              {
-                name: `Additional Info`,
-                value: `${additionalInput}`,
-                inline: false,
-              },
-              {
-                name: `Type`,
-                value: `${data.Ticket}`,
-                inline: false,
-              },
-            ])
-            .setFooter({ text: `${interaction.guild.name} Tickets` })
-            .setTimestamp()
-            .setImage("https://femboy.kz/images/wide.png");
-
-          const buttonClose = new ActionRowBuilder().addComponents(
-            new ButtonBuilder()
-              .setCustomId("ticket-close")
-              .setLabel("Close ticket")
-              .setStyle(ButtonStyle.Danger)
-          );
-
-          let channel = await interaction.guild.channels.create({
-            name: `ticket-${interaction.user.id}`,
-            type: ChannelType.GuildText,
-            parent: `${category}`,
-          });
-
-          let msg = await channel.send({
-            embeds: [embed],
-            components: [buttonClose],
-          });
-          await interaction.reply({
-            content: `You have opened a ticket: ${channel}`,
-            ephemeral: true,
-          });
-
-          const member = interaction.user.id;
-          await channel.permissionOverwrites.edit(member, {
-            SendMessages: true,
-            ViewChannel: true,
-          });
-
-          const collector = msg.createMessageComponentCollector();
-
-          collector.on("collect", async (i) => {
-            if (
-              !i.member.permissions.has(PermissionsBitField.Flags.Administrator)
-            )
-              return await interaction.reply({
-                content: "You don't have perms to use this command.",
-                ephemeral: true,
-              });
-
-            await channel.delete();
-          });
-        });
-    }
-  }
-});
-
-// CS:GO & Mc whitelist request commands, unban request command, report command
-const whitelistSchema = require("./Schemas/whitelistSchema");
-const mcWhitelistSchema = require("./Schemas/mcWhitelistSchema");
-const unbanSchema = require("./Schemas/unbanSchema");
-const reportSchema = require("./Schemas/reportSchema");
-
-client.on(Events.InteractionCreate, async (interaction) => {
-  if (!interaction.isModalSubmit()) return;
-
-  if (interaction.customId === "modalMc") {
-    const uuid = interaction.fields.getTextInputValue("uuidMc");
-    const reason = interaction.fields.getTextInputValue("reasonMc");
-    const request = interaction.fields.getTextInputValue("requestMc");
-
-    const member = interaction.user.id;
-    const tag = interaction.user.tag;
-    const server = interaction.guild.name;
-    const serverId = interaction.guild.id;
-
-    if (reason.length > 500 || request.length > 500) {
-      return await interaction.reply({
-        content: `You have entered too much text, please shorten it and try again.`,
-        ephemeral: true,
-      });
-    }
-
-    const embedMc = new EmbedBuilder()
-      .setColor("#ff00b3")
-      .setTitle("New Whitelist Request")
-      .setImage("https://femboy.kz/images/wide.png")
-      .setDescription(
-        `Requesting member: ${tag} (${member})\nIn Server: ${server} (${serverId})`
-      )
-      .addFields({
-        name: "Name / UUID",
-        value: `${uuid}`,
-        inline: false,
-      })
-      .addFields({
-        name: "Reasoning for whitelist",
-        value: `${reason}`,
-        inline: false,
-      })
-      .addFields({
-        name: "What they want to do on the server",
-        value: `${request}`,
-        inline: false,
-      })
-      .setTimestamp();
-
-    mcWhitelistSchema
-      .findOne({ Guild: interaction.guild.id })
-      .then(async (data) => {
-        if (!data) return;
-
-        const channelID = data.Channel;
-        const channel = interaction.guild.channels.cache.get(channelID);
-
-        channel.send({ embeds: [embedMc] });
-
-        await interaction.reply({
-          content: "Your request has been submitted.",
-          ephemeral: true,
-        });
-      })
-      .catch(async (error) => {
-        console.error("Error executing command:", error);
-        await interaction.reply({
-          content: "There was an error while executing this command!",
-          ephemeral: true,
-        });
-      });
-  }
-
-  if (interaction.customId === "modalWhitelist") {
-    const steam = interaction.fields.getTextInputValue("steamWhitelist");
-    const reason = interaction.fields.getTextInputValue("reasonWhitelist");
-    const request = interaction.fields.getTextInputValue("requestWhitelist");
-
-    const member = interaction.user.id;
-    const tag = interaction.user.tag;
-    const guild = interaction.guild.name;
-    const guildId = interaction.guild.id;
-
-    if (reason.length > 500 || request.length > 500) {
-      return await interaction.reply({
-        content: `You have entered too much text, please shorten it and try again.`,
-        ephemeral: true,
-      });
-    }
-
-    const embedWhitelist = new EmbedBuilder()
-      .setColor("#ff00b3")
-      .setTitle("New Whitelist Request")
-      .setImage("https://femboy.kz/images/wide.png")
-      .setDescription(
-        `Requesting member: ${tag} (${member})\nIn Server: ${guild} (${guildId})`
-      )
-      .addFields({
-        name: "SteamID / Profile URL",
-        value: `${steam}`,
-        inline: false,
-      })
-      .addFields({
-        name: "Reasoning for whitelist",
-        value: `${reason}`,
-        inline: false,
-      })
-      .addFields({
-        name: "Requested to Group? Yes/No",
-        value: `${request}`,
-        inline: false,
-      })
-      .setTimestamp();
-
-    whitelistSchema
-      .findOne({ Guild: interaction.guild.id })
-      .then(async (data) => {
-        if (!data) return;
-
-        const channelID = data.Channel;
-        const channel = interaction.guild.channels.cache.get(channelID);
-
-        channel.send({ embeds: [embedWhitelist] });
-
-        await interaction.reply({
-          content: "Your request has been submitted.",
-          ephemeral: true,
-        });
-      })
-      .catch(async (error) => {
-        console.error("Error executing command:", error);
-        await interaction.reply({
-          content: "There was an error while executing this command!",
-          ephemeral: true,
-        });
-      });
-  }
-
-  // unban request command
-  if (interaction.customId === "modalUnban") {
-    const steam = interaction.fields.getTextInputValue("steamUnban");
-    const reason = interaction.fields.getTextInputValue("reasonUnban");
-    const server = interaction.fields.getTextInputValue("serverUnban");
-
-    const member = interaction.user.id;
-    const tag = interaction.user.tag;
-    const guild = interaction.guild.name;
-    const guildId = interaction.guild.id;
-
-    if (reason.length > 500) {
-      return await interaction.reply({
-        content: `You have entered too much text, please shorten it and try again.`,
-        ephemeral: true,
-      });
-    }
-
-    const embedUnban = new EmbedBuilder()
-      .setColor("#ff00b3")
-      .setTitle("New Unban Request")
-      .setImage("https://femboy.kz/images/wide.png")
-      .setDescription(
-        `Requesting member: ${tag} (${member})\nIn Server: ${guild} (${guildId})`
-      )
-      .addFields({
-        name: "SteamID / Profile URL",
-        value: `${steam}`,
-        inline: false,
-      })
-      .addFields({
-        name: "Reasoning for unban",
-        value: `${reason}`,
-        inline: false,
-      })
-      .addFields({
-        name: "Server IP/Name",
-        value: `${server}`,
-        inline: false,
-      })
-      .setTimestamp();
-
-    unbanSchema
-      .findOne({ Guild: interaction.guild.id })
-      .then(async (data) => {
-        if (!data) return;
-
-        const channelID = data.Channel;
-        const channel = interaction.guild.channels.cache.get(channelID);
-
-        channel.send({ embeds: [embedUnban] });
-
-        await interaction.reply({
-          content: "Your request has been submitted.",
-          ephemeral: true,
-        });
-      })
-      .catch(async (error) => {
-        console.error("Error executing command:", error);
-        await interaction.reply({
-          content: "There was an error while executing this command!",
-          ephemeral: true,
-        });
-      });
-  }
-
-  // report/suggestion command
-  if (interaction.customId === "modalReport") {
-    const issue = interaction.fields.getTextInputValue("issueReport");
-    const info = interaction.fields.getTextInputValue("infoReport");
-    const more = interaction.fields.getTextInputValue("moreReport");
-
-    const member = interaction.user.id;
-    const tag = interaction.user.tag;
-    const guild = interaction.guild.name;
-    const guildId = interaction.guild.id;
-
-    if (info.length > 500 || more.length > 500) {
-      return await interaction.reply({
-        content: `You have entered too much text, please shorten it and try again.`,
-        ephemeral: true,
-      });
-    }
-
-    const embedReport = new EmbedBuilder()
-      .setColor("Red")
-      .setTitle("New Report/Suggestion Request")
-      .setImage("https://femboy.kz/images/wide.png")
-      .setDescription(
-        `Requesting member: ${tag} (${member})\nIn Server: ${guild} (${guildId})`
-      )
-      .addFields({
-        name: "Reported issue / Suggestion",
-        value: `${issue}`,
-        inline: false,
-      })
-      .addFields({
-        name: "Issue/Suggestion in detail",
-        value: `${info}`,
-        inline: false,
-      })
-      .addFields({
-        name: "More info, such as links.",
-        value: `${more}`,
-        inline: false,
-      })
-      .setTimestamp();
-
-    reportSchema
-      .findOne({ Guild: interaction.guild.id })
-      .then(async (data) => {
-        if (!data) return;
-
-        const channelID = data.Channel;
-        const channel = interaction.guild.channels.cache.get(channelID);
-
-        channel.send({ embeds: [embedReport] });
-
-        await interaction.reply({
-          content: "Your request has been submitted.",
-          ephemeral: true,
-        });
-      })
-      .catch(async (error) => {
-        console.error("Error executing command:", error);
-        await interaction.reply({
-          content: "There was an error while executing this command!",
-          ephemeral: true,
-        });
-      });
-  }
-});
-
-// antilink system
-const linkSchema = require("./Schemas/linkSchema");
-
-client.on(Events.MessageCreate, async (message) => {
-  if (
-    message.content.startsWith("http") ||
-    message.content.startsWith("discord.gg") ||
-    message.content.includes("https://") ||
-    message.content.includes("http://") ||
-    message.content.includes("discord.gg/")
-  ) {
-    const Data = await linkSchema.findOne({ Guild: message.guild.id });
-
-    if (!Data) return;
-
-    const memberPerms = Data.Perms;
-    const user = message.author;
-    const member = message.guild.members.cache.get(user.id);
-
-    if (member.permissions.has(memberPerms)) return;
-    else {
-      await (
-        await message.channel.send({
-          conent: `${message.author}, you can't send links here!`,
-        })
-      ).then((msg) => {
-        setTimeout(() => msg.delete(), 3000);
-      });
-      (await message).delete();
-    }
-  }
-});
 
 // ------------------------------------------------------------------------ guild scraper
 async () => {
@@ -634,83 +148,6 @@ client.on(Events.InteractionCreate, async (interaction) => {
   }
 });
 
-// bot join/leave logger, useless feature
-// doesn't have an enable/disable command and only runs on fkz server cuz lazy lol
-client.on(Events.GuildCreate, async (guild) => {
-  const channel = await client.channels.cache.get(
-    `${process.env.LOGS_CHAT_ID}`
-  );
-  const name = guild.name;
-  const serverID = guild.id;
-  const memberCount = guild.memberCount;
-
-  const ownerID = guild.ownerId;
-  const owner = await client.users.cache.get(ownerID);
-  const ownerName = owner.username;
-
-  const embed = new EmbedBuilder()
-    .setColor("Green")
-    .setTitle("FKZ Bot has joined a new server.")
-    .addFields([
-      {
-        name: "Server",
-        value: `> ${name} / ${serverID}`,
-      },
-      {
-        name: "Server Membercount",
-        value: `> ${memberCount}`,
-      },
-      {
-        name: "Server Owner",
-        value: `> ${ownerName} / ${ownerID}`,
-      },
-      {
-        name: "Server Age",
-        value: `> <t:${parseInt(guild.createdTimestamp / 1000)}:R>`,
-      },
-    ])
-    .setTimestamp()
-    .setFooter({ text: "Guild Join" });
-  await channel.send({ embeds: [embed] });
-});
-client.on(Events.GuildDelete, async (guild) => {
-  const channel = await client.channels.cache.get(
-    `${process.env.LOGS_CHAT_ID}`
-  );
-  const name = guild.name;
-  const serverID = guild.id;
-  const memberCount = guild.memberCount;
-
-  const ownerID = guild.ownerId;
-  const owner = await client.users.cache.get(ownerID);
-  const ownerName = owner.username;
-
-  const embed = new EmbedBuilder()
-    .setColor("Red")
-    .setTitle("FKZ Bot has left a server.")
-    .addFields([
-      {
-        name: "Server",
-        value: `> ${name} / ${serverID}`,
-      },
-      {
-        name: "Server Membercount",
-        value: `> ${memberCount}`,
-      },
-      {
-        name: "Server Owner",
-        value: `> ${ownerName} / ${ownerID}`,
-      },
-      {
-        name: "Server Age",
-        value: `> <t:${parseInt(guild.createdTimestamp / 1000)}:R>`,
-      },
-    ])
-    .setTimestamp()
-    .setFooter({ text: "Guild Leave" });
-  await channel.send({ embeds: [embed] });
-});
-
 // REACTION ROLES //
 const reactions = require("./Schemas/reactionrs");
 
@@ -763,1468 +200,7 @@ client.on(Events.MessageReactionRemove, async (reaction, user) => {
   }
 });
 
-// AUDIT LOG //
-const Audit_Log = require("./Schemas/auditlog");
-// ------------------------------------------------------------------------ ban logs
-client.on(Events.GuildBanAdd, async (guild, user) => {
-  const data = await Audit_Log.findOne({
-    Guild: guild.id,
-  });
-  let logID;
-  if (data) {
-    logID = data.Channel;
-  } else {
-    return;
-  }
-  const banInfo = await guild.fetchBan(user);
-  if (!banInfo) return;
-
-  const { reason, executor } = banInfo;
-  const auditChannel = client.channels.cache.get(logID);
-
-  const auditEmbed = new EmbedBuilder()
-    .setColor("#ff00b3")
-    .setTimestamp()
-    .setFooter({ text: "FKZ Log System" })
-    .setTitle("Ban Added")
-    .addFields(
-      {
-        name: "Banned Member:",
-        value: `${user.tag || "unknown"}`,
-        inline: false,
-      },
-      {
-        name: "Executor:",
-        value: `${executor.tag || "unknown"}`,
-        inline: false,
-      },
-      {
-        name: "Reason:",
-        value: `${reason || "none"}`,
-        inline: false,
-      }
-    );
-  await auditChannel.send({ embeds: [auditEmbed] });
-});
-client.on(Events.GuildBanRemove, async (user) => {
-  const data = await Audit_Log.findOne({
-    Guild: user.guild.id,
-  });
-  let logID;
-  if (data) {
-    logID = data.Channel;
-  } else {
-    return;
-  }
-  const banInfo = await user.guild.fetchBan(user);
-  if (!banInfo) return;
-
-  const { executor } = banInfo;
-
-  const auditChannel = client.channels.cache.get(logID);
-
-  user.fetch().then(async (fullUser) => {
-    const auditEmbed = new EmbedBuilder()
-      .setColor("#ff00b3")
-      .setTimestamp()
-      .setFooter({ text: "FKZ Log System" })
-      .setTitle("Ban Removed")
-      .addFields(
-        {
-          name: "Member:",
-          value: `${fullUser.user || user.user || "unknown"}`,
-          inline: false,
-        },
-        {
-          name: "Admin:",
-          value: `${executor || "unknown"}`,
-          inline: false,
-        }
-      );
-    await auditChannel.send({ embeds: [auditEmbed] });
-  });
-});
-// ------------------------------------------------------------------------ channel logs
-client.on(Events.ChannelCreate, async (channel) => {
-  const data = await Audit_Log.findOne({
-    Guild: channel.guild.id,
-  });
-  let logID;
-  if (data) {
-    logID = data.Channel;
-  } else {
-    return;
-  }
-  const auditChannel = client.channels.cache.get(logID);
-
-  const auditEmbed = new EmbedBuilder()
-    .setColor("#ff00b3")
-    .setTimestamp()
-    .setFooter({ text: "FKZ Log System" })
-    .setTitle("Channel Created")
-    .addFields(
-      {
-        name: "Name:",
-        value: `${channel.name || "?"}`,
-        inline: false,
-      },
-      {
-        name: "Type:",
-        value: `${channel.type}`,
-        inline: false,
-      },
-      {
-        name: "Category:",
-        value: `${channel.parent || "No Category"}`,
-        inline: false,
-      },
-      {
-        name: "ID:",
-        value: `${channel.id}`,
-        inline: false,
-      }
-    );
-  await auditChannel.send({ embeds: [auditEmbed] });
-});
-client.on(Events.ChannelDelete, async (channel) => {
-  const data = await Audit_Log.findOne({
-    Guild: channel.guild.id,
-  });
-  let logID;
-  if (data) {
-    logID = data.Channel;
-  } else {
-    return;
-  }
-  const auditChannel = client.channels.cache.get(logID);
-
-  const auditEmbed = new EmbedBuilder()
-    .setColor("#ff00b3")
-    .setTimestamp()
-    .setFooter({ text: "FKZ Log System" })
-    .setTitle("Channel Deleted")
-    .addFields(
-      {
-        name: "Name:",
-        value: `${channel.name || "?"}`,
-        inline: false,
-      },
-      {
-        name: "Type:",
-        value: `${channel.type}`,
-        inline: false,
-      },
-      {
-        name: "Category:",
-        value: `${channel.parent || "No Category"}`,
-        inline: false,
-      },
-      {
-        name: "ID:",
-        value: `${channel.id}`,
-        inline: false,
-      }
-    );
-  await auditChannel.send({ embeds: [auditEmbed] });
-});
-client.on(Events.ChannelUpdate, async (oldChannel, newChannel) => {
-  const data = await Audit_Log.findOne({
-    Guild: oldChannel.guild.id,
-  });
-  let logID;
-  if (data) {
-    logID = data.Channel;
-  } else {
-    return;
-  }
-  const auditChannel = client.channels.cache.get(logID);
-  const changes = [];
-
-  if (oldChannel.name !== newChannel.name) {
-    changes.push(
-      `Name: \`${oldChannel.name || "none"}\` → \`${
-        newChannel.name || "none"
-      }\``
-    );
-  }
-  if (oldChannel.parent !== newChannel.parent) {
-    changes.push(
-      `Category: \`${oldChannel.parent || "none"}\` → \`${
-        newChannel.parent || "none"
-      }\``
-    );
-  }
-  if (oldChannel.topic !== newChannel.topic) {
-    changes.push(
-      `Topic: \`${oldChannel.topic || "None"}\` → \`${
-        newChannel.topic || "None"
-      }\``
-    );
-  }
-
-  if (changes.length === 0) return;
-  const changesText = changes.join("\n");
-
-  const auditEmbed = new EmbedBuilder()
-    .setColor("#ff00b3")
-    .setTimestamp()
-    .setFooter({ text: "FKZ Log System" })
-    .setTitle("Channel Updated")
-    .addFields(
-      { name: "Changes:", value: `${changesText}`, inline: false },
-      { name: "ID:", value: `${newChannel.id}`, inline: false }
-    );
-  await auditChannel.send({ embeds: [auditEmbed] });
-});
-// ------------------------------------------------------------------------ Invite logs
-client.on(Events.InviteCreate, async (invite) => {
-  const data = await Audit_Log.findOne({
-    Guild: invite.guild.id,
-  });
-  let logID;
-  if (data) {
-    logID = data.Channel;
-  } else {
-    return;
-  }
-  const auditChannel = client.channels.cache.get(logID);
-
-  const auditEmbed = new EmbedBuilder()
-    .setColor("#ff00b3")
-    .setTimestamp()
-    .setFooter({ text: "FKZ Log System" })
-    .setTitle("Invite Created")
-    .addFields(
-      {
-        name: "Creator:",
-        value: `${invite.inviterId || "unknown"}`,
-        inline: false,
-      },
-      {
-        name: "Url:",
-        value: `${invite.url}`,
-        inline: false,
-      }
-    );
-  await auditChannel.send({ embeds: [auditEmbed] });
-});
-client.on(Events.InviteDelete, async (invite) => {
-  const data = await Audit_Log.findOne({
-    Guild: invite.guild.id,
-  });
-  let logID;
-  if (data) {
-    logID = data.Channel;
-  } else {
-    return;
-  }
-  const auditChannel = client.channels.cache.get(logID);
-
-  const auditEmbed = new EmbedBuilder()
-    .setColor("#ff00b3")
-    .setTimestamp()
-    .setFooter({ text: "FKZ Log System" })
-    .setTitle("Invite Deleted")
-    .addFields(
-      {
-        name: "Author:",
-        value: `${invite.inviter || "unknown"}`,
-        inline: false,
-      },
-      {
-        name: "Uses / Max Uses:",
-        value: `${invite.uses || "unknown"} / ${invite.maxUses || "unknown"}`,
-        inline: false,
-      },
-      {
-        name: "Channel:",
-        value: `${invite.channel || "unknown"}`,
-        inline: false,
-      },
-      {
-        name: "Invite:",
-        value: `${invite.code}`,
-        inline: false,
-      },
-      {
-        name: "Url:",
-        value: `${invite.url}`,
-        inline: false,
-      }
-    );
-  await auditChannel.send({ embeds: [auditEmbed] });
-});
-// ------------------------------------------------------------------------ emoji logs
-client.on(Events.GuildEmojiCreate, async (emoji) => {
-  const data = await Audit_Log.findOne({
-    Guild: emoji.guild.id,
-  });
-  let logID;
-  if (data) {
-    logID = data.Channel;
-  } else {
-    return;
-  }
-  const auditChannel = client.channels.cache.get(logID);
-  let image = emoji.imageURL({ size: 64 });
-
-  const auditEmbed = new EmbedBuilder()
-    .setColor("#ff00b3")
-    .setTimestamp()
-    .setImage(`${image}` || `https://femboy.kz/images/wide.png`)
-    .setFooter({ text: "FKZ Log System" })
-    .setTitle("Emoji Added")
-    .addFields(
-      { name: "Name:", value: `${emoji.name || "?"}`, inline: false },
-      { name: "Author:", value: `${emoji.author || "unknown"}`, inline: false },
-      { name: "Animated?:", value: `${emoji.animated || "?"}`, inline: false },
-      { name: "ID:", value: `${emoji.id}`, inline: false }
-    );
-  await auditChannel.send({ embeds: [auditEmbed] });
-});
-client.on(Events.GuildEmojiDelete, async (emoji) => {
-  const data = await Audit_Log.findOne({
-    Guild: emoji.guild.id,
-  });
-  let logID;
-  if (data) {
-    logID = data.Channel;
-  } else {
-    return;
-  }
-  const auditChannel = client.channels.cache.get(logID);
-  let image = emoji.imageURL({ size: 64 });
-
-  const auditEmbed = new EmbedBuilder()
-    .setColor("#ff00b3")
-    .setTimestamp()
-    .setImage(`${image}` || `https://femboy.kz/images/wide.png`)
-    .setFooter({ text: "FKZ Log System" })
-    .setTitle("Emoji Deleted")
-    .addFields(
-      { name: "Name:", value: `${emoji.name || "?"}`, inline: false },
-      { name: "Author:", value: `${emoji.author || "unknown"}`, inline: false },
-      { name: "Animated?:", value: `${emoji.animated || "?"}`, inline: false },
-      { name: "ID:", value: `${emoji.id}`, inline: false }
-    );
-  await auditChannel.send({ embeds: [auditEmbed] });
-});
-client.on(Events.GuildEmojiUpdate, async (oldEmoji, newEmoji) => {
-  const data = await Audit_Log.findOne({
-    Guild: newEmoji.guild.id,
-  });
-  let logID;
-  if (data) {
-    logID = data.Channel;
-  } else {
-    return;
-  }
-  const auditChannel = client.channels.cache.get(logID);
-  const changes = [];
-
-  if (oldEmoji.name !== newEmoji.name) {
-    changes.push(
-      `Name: \`${oldEmoji.name || "none"}\` → \`${newEmoji.name || "none"}\``
-    );
-  }
-
-  if (changes.length === 0) return;
-  const changesText = changes.join("\n");
-  let image = newEmoji.imageURL({ size: 64 });
-
-  const auditEmbed = new EmbedBuilder()
-    .setColor("#ff00b3")
-    .setTimestamp()
-    .setImage(`${image}` || `https://femboy.kz/images/wide.png`)
-    .setFooter({ text: "FKZ Log System" })
-    .setTitle("Emoji Edited")
-    .addFields(
-      { name: "Changes:", value: `${changesText}`, inline: false },
-      { name: "Author:", value: `${newEmoji.author || "null"}`, inline: false },
-      { name: "ID:", value: `${newEmoji.id}`, inline: false }
-    );
-  await auditChannel.send({ embeds: [auditEmbed] });
-});
-// ------------------------------------------------------------------------ Sticker logs
-client.on(Events.GuildStickerCreate, async (sticker) => {
-  const data = await Audit_Log.findOne({
-    Guild: sticker.guild.id,
-  });
-  let logID;
-  if (data) {
-    logID = data.Channel;
-  } else {
-    return;
-  }
-  const auditChannel = client.channels.cache.get(logID);
-  let image = sticker.imageURL({ size: 64 });
-
-  const auditEmbed = new EmbedBuilder()
-    .setColor("#ff00b3")
-    .setTimestamp()
-    .setImage(`${image}` || `https://femboy.kz/images/wide.png`)
-    .setFooter({ text: "FKZ Log System" })
-    .setTitle("Sticker Added")
-    .addFields(
-      {
-        name: "Name:",
-        value: `${sticker.name}`,
-        inline: false,
-      },
-      {
-        name: "Description:",
-        value: `${sticker.description || "none"}`,
-        inline: false,
-      },
-      {
-        name: "Format:",
-        value: `${sticker.format || "none"}`,
-        inline: false,
-      },
-      {
-        name: "ID:",
-        value: `${sticker.id}`,
-        inline: false,
-      }
-    );
-  await auditChannel.send({ embeds: [auditEmbed] });
-});
-client.on(Events.GuildStickerDelete, async (sticker) => {
-  const data = await Audit_Log.findOne({
-    Guild: sticker.guild.id,
-  });
-  let logID;
-  if (data) {
-    logID = data.Channel;
-  } else {
-    return;
-  }
-  const auditChannel = client.channels.cache.get(logID);
-  let image = sticker.url({ size: 128 });
-
-  const auditEmbed = new EmbedBuilder()
-    .setColor("#ff00b3")
-    .setTimestamp()
-    .setImage(`${image}` || `https://femboy.kz/images/wide.png`)
-    .setFooter({ text: "FKZ Log System" })
-    .setTitle("Sticker Deleted")
-    .addFields(
-      { name: "Name:", value: `${sticker.name}`, inline: false },
-      { name: "ID:", value: `${sticker.id}`, inline: false }
-    );
-  await auditChannel.send({ embeds: [auditEmbed] });
-});
-client.on(Events.GuildStickerUpdate, async (oldSticker, newSticker) => {
-  const data = await Audit_Log.findOne({
-    Guild: newSticker.guild.id,
-  });
-  let logID;
-  if (data) {
-    logID = data.Channel;
-  } else {
-    return;
-  }
-  const auditChannel = client.channels.cache.get(logID);
-  const changes = [];
-
-  if (oldSticker.name !== newSticker.name) {
-    changes.push(`Name: \`${oldSticker.name}\` → \`${newSticker.name}\``);
-  }
-  if (oldSticker.description !== newSticker.description) {
-    changes.push(
-      `Description: \`${oldSticker.description || "None"}\` → \`${
-        newSticker.description || "None"
-      }\``
-    );
-  }
-
-  if (changes.length === 0) return;
-  const changesText = changes.join("\n");
-  let image = newSticker.url({ size: 128 });
-
-  const auditEmbed = new EmbedBuilder()
-    .setColor("#ff00b3")
-    .setTimestamp()
-    .setImage(`${image}` || `https://femboy.kz/images/wide.png`)
-    .setFooter({ text: "FKZ Log System" })
-    .setTitle("Sticker Edited")
-    .addFields(
-      { name: "Changes:", value: `${changesText}`, inline: false },
-      { name: "ID:", value: `${newSticker.id}`, inline: false }
-    );
-  await auditChannel.send({ embeds: [auditEmbed] });
-});
-// ------------------------------------------------------------------------ Role logs
-client.on(Events.GuildRoleCreate, async (role) => {
-  const data = await Audit_Log.findOne({
-    Guild: role.guild.id,
-  });
-  let logID;
-  if (data) {
-    logID = data.Channel;
-  } else {
-    return;
-  }
-  const auditChannel = client.channels.cache.get(logID);
-
-  const auditEmbed = new EmbedBuilder()
-    .setColor("#ff00b3")
-    .setTimestamp()
-    .setFooter({ text: "FKZ Log System" })
-    .setTitle("Role Created")
-    .addFields(
-      { name: "Name:", value: `${role.name}`, inline: false },
-      { name: "ID:", value: `${role.id}`, inline: false }
-    );
-  await auditChannel.send({ embeds: [auditEmbed] });
-});
-client.on(Events.GuildRoleDelete, async (role) => {
-  const data = await Audit_Log.findOne({
-    Guild: role.guild.id,
-  });
-  let logID;
-  if (data) {
-    logID = data.Channel;
-  } else {
-    return;
-  }
-  const auditChannel = client.channels.cache.get(logID);
-
-  const auditEmbed = new EmbedBuilder()
-    .setColor("#ff00b3")
-    .setTimestamp()
-    .setFooter({ text: "FKZ Log System" })
-    .setTitle("Role Deleted")
-    .addFields(
-      { name: "Name:", value: `${role.name}`, inline: false },
-      { name: "ID:", value: `${role.id}`, inline: false }
-    );
-  await auditChannel.send({ embeds: [auditEmbed] });
-});
-client.on(Events.GuildRoleUpdate, async (oldRole, newRole) => {
-  const data = await Audit_Log.findOne({
-    Guild: newRole.guild.id,
-  });
-  let logID;
-  if (data) {
-    logID = data.Channel;
-  } else {
-    return;
-  }
-  const auditChannel = client.channels.cache.get(logID);
-  const changes = [];
-
-  if (oldRole.name !== newRole.name) {
-    changes.push(`Name: \`${oldRole.name}\` → \`${newRole.name}\``);
-  }
-
-  if (oldRole.permissions !== newRole.permissions) {
-    changes.push(
-      `Permissions: \`${oldRole.permissions || "None"}\` → \`${
-        newRole.permissions || "None"
-      }\``
-    );
-  }
-
-  if (oldRole.mentionable !== newRole.mentionable) {
-    changes.push(
-      `Mentionable?: \`${oldRole.mentionable || "None"}\` → \`${
-        newRole.mentionable || "None"
-      }\``
-    );
-  }
-
-  if (oldRole.hoist !== newRole.hoist) {
-    changes.push(
-      `Hoisted?: \`${oldRole.hoist || "None"}\` → \`${
-        newRole.hoist || "None"
-      }\``
-    );
-  }
-
-  if (oldRole.color !== newRole.color) {
-    changes.push(
-      `Color: \`${oldRole.color + " - " + oldRole.hexColor || "None"}\` → \`${
-        newRole.color + " - " + newRole.hexColor || "None"
-      }\``
-    );
-  }
-
-  if (changes.length === 0) return;
-  const changesText = changes.join("\n");
-
-  const auditEmbed = new EmbedBuilder()
-    .setColor("#ff00b3")
-    .setTimestamp()
-    .setTitle("Role Updated")
-    .setFooter({ text: "FKZ Log System" })
-    .addFields(
-      { name: "Changes:", value: `${changesText}`, inline: false },
-      { name: "ID:", value: `${newRole.id}`, inline: false }
-    );
-  await auditChannel.send({ embeds: [auditEmbed] });
-});
-// ------------------------------------------------------------------------ message logs
-client.on(Events.MessageDelete, async (message) => {
-  const data = await Audit_Log.findOne({
-    Guild: message.guild.id,
-  });
-  let logID;
-  if (data) {
-    logID = data.Channel;
-  } else {
-    return;
-  }
-  if (!message.author) return;
-  if (message.author.bot) return;
-  if (message.author.id !== client.user.id) return;
-  if (message.partial) console.log("deleted msg is partial");
-
-  try {
-    message.fetch().then(async (fullMessage) => {
-      const auditChannel = client.channels.cache.get(logID);
-
-      if (message.content.length >= 3072) return;
-      if (fullMessage.content.length >= 3072) return;
-
-      const auditEmbed = new EmbedBuilder()
-        .setColor("#ff00b3")
-        .setTimestamp()
-        .setFooter({ text: "FKZ Log System" })
-        .setTitle("Message Deleted")
-        .setAuthor({ name: "Message:" })
-        .setDescription(`${fullMessage.content || message.content || "none"}`)
-        .addFields(
-          {
-            name: "Author:",
-            value: `${fullMessage.author || message.author}`,
-            inline: false,
-          },
-          {
-            name: "Channel:",
-            value: `${fullMessage.channel || message.channel}`,
-            inline: false,
-          },
-          {
-            name: "MessageID:",
-            value: `${fullMessage.id || message.id}`,
-            inline: false,
-          }
-        );
-      await auditChannel.send({ embeds: [auditEmbed] });
-    });
-  } catch (err) {
-    return console.log(err);
-  }
-});
-client.on(Events.MessageUpdate, async (oldMessage, newMessage) => {
-  const data = await Audit_Log.findOne({
-    Guild: newMessage.guild.id,
-  });
-  let logID;
-  if (data) {
-    logID = data.Channel;
-  } else {
-    return;
-  }
-  if (!oldMessage.author) return;
-  if (oldMessage.author.bot) return;
-  if (oldMessage.partial) console.log("oldMessage = partial");
-
-  try {
-    const auditChannel = client.channels.cache.get(logID);
-    const changes = [];
-
-    if (oldMessage.content !== newMessage.content) {
-      oldMessage.fetch().then((fullOldMessage) => {
-        const fullOldMessageText = fullOldMessage.content;
-        const oldMessageText = oldMessage.content;
-        const fullOldAuthor = fullOldMessage.author;
-        changes.push(
-          `Message: \`${fullOldMessageText || oldMessageText || "None"}\` → \`${
-            newMessage.content || "None"
-          }\``
-        );
-
-        if (fullOldMessageText || oldMessageText >= 1536) return;
-        if (newMessage.content.length >= 1536) return;
-        if (changes.length === 0) return;
-        const changesText = changes.join("\n");
-
-        const auditEmbed = new EmbedBuilder()
-          .setColor("#ff00b3")
-          .setTimestamp()
-          .setFooter({ text: "FKZ Log System" })
-          .setTitle("Message Edited")
-          .setAuthor({ name: "Edit:" })
-          .setDescription(`${changesText}`)
-          .addFields(
-            {
-              name: "Author:",
-              value: `${fullOldAuthor || oldMessage.author || "unknown"}`,
-              inline: false,
-            },
-            {
-              name: "Channel:",
-              value: `${newMessage.channel}`,
-              inline: false,
-            },
-            {
-              name: "MessageID:",
-              value: `${newMessage.id}`,
-            }
-          );
-        auditChannel.send({ embeds: [auditEmbed] });
-      });
-    }
-  } catch (err) {
-    return console.log(err);
-  }
-});
-// ------------------------------------------------------------------------ Automod logs
-client.on(Events.AutoModerationRuleCreate, async (autoModerationRule) => {
-  const data = await Audit_Log.findOne({
-    Guild: autoModerationRule.guild.id,
-  });
-  let logID;
-  if (data) {
-    logID = data.Channel;
-  } else {
-    return;
-  }
-
-  const auditChannel = client.channels.cache.get(logID);
-  let actions = autoModerationRule.actions.toString();
-
-  const auditEmbed = new EmbedBuilder()
-    .setColor("#ff00b3")
-    .setTimestamp()
-    .setFooter({ text: "FKZ Log System" })
-    .setTitle("Automod Rule Created")
-    .addFields(
-      {
-        name: "Creator:",
-        value: `<@${autoModerationRule.creatorId}>`,
-        inline: false,
-      },
-      {
-        name: "Name:",
-        value: `${autoModerationRule.name}`,
-        inline: false,
-      },
-      {
-        name: "Actions:",
-        value: `${actions}`,
-        inline: false,
-      }
-    );
-  await auditChannel.send({ embeds: [auditEmbed] });
-});
-client.on(Events.AutoModerationRuleDelete, async (autoModerationRule) => {
-  const data = await Audit_Log.findOne({
-    Guild: autoModerationRule.guild.id,
-  });
-  let logID;
-  if (data) {
-    logID = data.Channel;
-  } else {
-    return;
-  }
-  const auditChannel = client.channels.cache.get(logID);
-  let actions = autoModerationRule.actions.toString();
-
-  const auditEmbed = new EmbedBuilder()
-    .setColor("#ff00b3")
-    .setTimestamp()
-    .setFooter({ text: "FKZ Log System" })
-    .setTitle("Automod Rule Deleted")
-    .addFields(
-      {
-        name: "Creator:",
-        value: `<@${autoModerationRule.creatorId}>`,
-        inline: false,
-      },
-      {
-        name: "Name:",
-        value: `${autoModerationRule.name}`,
-        inline: false,
-      },
-      {
-        name: "Actions:",
-        value: `${actions}`,
-        inline: false,
-      }
-    );
-  await auditChannel.send({ embeds: [auditEmbed] });
-});
-client.on(
-  Events.AutoModerationRuleUpdate,
-  // 1st param is "param or null" and reads null ???? --------------------------- #1
-  async (oldAutoModerationRule, newAutoModerationRule) => {
-    const data = await Audit_Log.findOne({
-      Guild: newAutoModerationRule.guild.id,
-    });
-    let logID;
-    if (data) {
-      logID = data.Channel;
-    } else {
-      return;
-    }
-
-    const auditChannel = client.channels.cache.get(logID);
-    const changes = [];
-
-    const nullEmbed = new EmbedBuilder()
-      .setColor("#ff00b3")
-      .setTimestamp()
-      .setFooter({ text: "FKZ Log System" })
-      .setTitle("Automod Rule Updated")
-      .addFields({
-        name: "Changes:",
-        value: `NULL`,
-      });
-
-    const auditEmbed = new EmbedBuilder()
-      .setColor("#ff00b3")
-      .setTimestamp()
-      .setFooter({ text: "FKZ Log System" })
-      .setTitle("Automod Rule Updated");
-
-    if (oldAutoModerationRule === null) {
-      return await auditChannel.send({ embeds: [nullEmbed] }); // avoid crash cuz of #1
-    }
-
-    if (oldAutoModerationRule.name !== newAutoModerationRule.name) {
-      changes.push(
-        `Name: \`${oldAutoModerationRule.name || "None"}\`  →  \`${
-          newAutoModerationRule.name || "None"
-        }\``
-      );
-      const changesText = changes.join("\n");
-      auditEmbed.addFields({
-        name: "Name Updated.",
-        value: `${changesText || "null"}`,
-      });
-      if (changes.length === 0) return;
-      await auditChannel.send({ embeds: [auditEmbed] });
-    }
-
-    if (oldAutoModerationRule.actions !== newAutoModerationRule.actions) {
-      let oldActions = oldAutoModerationRule.actions.toString();
-      let newActions = newAutoModerationRule.actions.toString();
-      auditEmbed.addFields(
-        {
-          name: "Old Rules:",
-          value: `${oldActions || "none"}`,
-          inline: false,
-        },
-        {
-          name: "New Rules:",
-          value: `${newActions || "none"}`,
-          inline: false,
-        }
-      );
-      if (oldAutoModerationRule.actions === null) return;
-      if (newAutoModerationRule.actions === null) return;
-      await auditChannel.send({ embeds: [auditEmbed] });
-    }
-
-    if (oldAutoModerationRule.enabled !== newAutoModerationRule.enabled) {
-      auditEmbed.addFields({
-        name: "Enabled?:",
-        value: `\`${oldAutoModerationRule.enabled || "Unknown"}\`  →  \`${
-          newAutoModerationRule.enabled || "Unknown"
-        }\``,
-        inline: false,
-      });
-      if (oldAutoModerationRule.enabled === null) return;
-      if (newAutoModerationRule.enabled === null) return;
-      await auditChannel.send({ embeds: [auditEmbed] });
-    }
-  }
-);
-// ------------------------------------------------------------------------ Thread logs
-client.on(Events.ThreadCreate, async (thread) => {
-  const data = await Audit_Log.findOne({
-    Guild: thread.guild.id,
-  });
-  let logID;
-  if (data) {
-    logID = data.Channel;
-  } else {
-    return;
-  }
-  const auditChannel = client.channels.cache.get(logID);
-  const auditEmbed = new EmbedBuilder()
-    .setColor("#ff00b3")
-    .setTimestamp()
-    .setFooter({ text: "FKZ Log System" })
-    .setTitle("Thread Created")
-    .addFields(
-      {
-        name: "Name:",
-        value: `${thread.name}`,
-        inline: false,
-      },
-      {
-        name: "Creator:",
-        value: `${thread.ownerId || "unknown"}`,
-        inline: false,
-      },
-      {
-        name: "Channel:",
-        value: `${thread.parent || "none"}`,
-        inline: false,
-      },
-      {
-        name: "Link:",
-        value: `<#${thread.id}>`,
-        inline: false,
-      },
-      {
-        name: "ID:",
-        value: `${thread.id}`,
-        inline: false,
-      }
-    );
-  await auditChannel.send({ embeds: [auditEmbed] });
-});
-client.on(Events.ThreadDelete, async (thread) => {
-  const data = await Audit_Log.findOne({
-    Guild: thread.guild.id,
-  });
-  let logID;
-  if (data) {
-    logID = data.Channel;
-  } else {
-    return;
-  }
-  const auditChannel = client.channels.cache.get(logID);
-
-  const auditEmbed = new EmbedBuilder()
-    .setColor("#ff00b3")
-    .setTimestamp()
-    .setFooter({ text: "FKZ Log System" })
-    .setTitle("Thread Deleted")
-    .addFields(
-      {
-        name: "Name:",
-        value: `${thread.name}`,
-        inline: false,
-      },
-      {
-        name: "Creator:",
-        value: `${thread.ownerId || "unknown"}`,
-        inline: false,
-      },
-      {
-        name: "Channel:",
-        value: `${thread.parent || "none"}`,
-        inline: false,
-      },
-      {
-        name: "Link:",
-        value: `<#${thread.id}>`,
-        inline: false,
-      },
-      {
-        name: "ID:",
-        value: `${thread.id}`,
-        inline: false,
-      }
-    );
-  await auditChannel.send({ embeds: [auditEmbed] });
-});
-client.on(Events.ThreadUpdate, async (oldThread, newThread) => {
-  const data = await Audit_Log.findOne({
-    Guild: oldThread.guild.id,
-  });
-  let logID;
-  if (data) {
-    logID = data.Channel;
-  } else {
-    return;
-  }
-
-  const auditChannel = client.channels.cache.get(logID);
-  const changes = [];
-
-  if (oldThread.name !== newThread.name) {
-    changes.push(`Name: \`${oldThread.name}\` → \`${newThread.name}\``);
-  }
-
-  if (oldThread.archived !== newThread.archived) {
-    changes.push(
-      `Archived?: \`${oldThread.archived || "None"}\` → \`${
-        newThread.archived || "None"
-      }\``
-    );
-  }
-  if (oldThread.locked !== newThread.locked) {
-    changes.push(
-      `Locked?: \`${oldThread.locked || "None"}\` → \`${
-        newThread.locked || "None"
-      }\``
-    );
-  }
-
-  if (changes.length === 0) return;
-  const changesText = changes.join("\n");
-
-  const auditEmbed = new EmbedBuilder()
-    .setColor("#ff00b3")
-    .setTimestamp()
-    .addFields(
-      {
-        name: "Creator:",
-        value: `${newThread.ownerId || "unknown"}`,
-        inline: false,
-      },
-      {
-        name: "Channel:",
-        value: `${newThread.parent || "unknown"}`,
-        inline: false,
-      },
-      {
-        name: "Link:",
-        value: `<#${newThread.id}>`,
-        inline: false,
-      },
-      {
-        name: "Changes:",
-        value: `${changesText}`,
-        inline: false,
-      },
-      {
-        name: "ID:",
-        value: `${newThread.id}`,
-        inline: false,
-      }
-    )
-    .setTitle("Thread Edited")
-    .setFooter({ text: "FKZ Log System" });
-  await auditChannel.send({ embeds: [auditEmbed] });
-});
-// ------------------------------------------------------------------------ member logs, invite join logs
-const invites = new Collection();
-const wait = require("timers/promises").setTimeout;
-client.on("ready", async () => {
-  await wait(2000);
-
-  client.guilds.cache.forEach(async (guild) => {
-    const clientMember = guild.members.cache.get(client.user.id);
-
-    if (!clientMember.permissions.has(PermissionsBitField.Flags.ManageGuild))
-      return;
-
-    const firstInvites = await guild.invites.fetch().catch((err) => {
-      console.log(err);
-    });
-
-    invites.set(
-      guild.id,
-      new Collection(firstInvites.map((invite) => [invite.code, invite.uses]))
-    );
-  });
-
-  client.on(Events.GuildMemberUpdate, async (oldMember, newMember) => {
-    const data = await Audit_Log.findOne({
-      Guild: oldMember.guild.id,
-    });
-    let logID;
-    if (data) {
-      logID = data.Channel;
-    } else {
-      return;
-    }
-
-    const logschatID = await client.channels.fetch(process.env.LOGS_CHAT_ID);
-
-    if (oldMember.user.system || newMember.user.system) return;
-    if (oldMember.partial) {
-      await oldMember.fetch().catch(() => {});
-    }
-
-    const auditEmbed = new EmbedBuilder()
-      .setColor("#ff00b3")
-      .setTimestamp()
-      .setFooter({ text: "FKZ Log System" })
-      .setTitle("Member Updated")
-      .addFields({
-        name: "User:",
-        value: `<@${newMember.user.id}> - ${newMember.user.username}`,
-        inline: false,
-      });
-
-    const auditChannel = client.channels.cache.get(logID);
-    if (auditChannel === undefined)
-      return console.log("undefined?", auditChannel);
-    const auditChannel2 = client.channels.cache.get(logschatID);
-    if (auditChannel2 === undefined)
-      return console.log("undefined? 2", auditChannel2);
-
-    if (oldMember.nickname !== newMember.nickname) {
-      oldMember.fetch().then(async (fullOldMember) => {
-        const fullOldNickName = fullOldMember.nickname;
-        auditEmbed.addFields({
-          name: `Nickname Updated`,
-          value: `DisplayName: \`${
-            fullOldNickName || oldMember.nickname || "none"
-          }\`  →  \`${newMember.nickname || "none"}\``,
-          inline: false,
-        });
-        if (fullOldNickName || oldMember.nickname === null)
-          return console.log("mem nickname null");
-        // await auditChannel.send({ embeds: [auditEmbed] });
-        await auditChannel2.send({ embeds: [auditEmbed] });
-      });
-    }
-
-    if (oldMember.displayName !== newMember.displayName) {
-      oldMember.fetch().then(async (fullOldMember) => {
-        const fullOldDisplayName = fullOldMember.displayName;
-        auditEmbed.addFields({
-          name: `Nickname or Displayname Updated`,
-          value: `DisplayName: \`${
-            fullOldDisplayName || oldMember.displayName || "none"
-          }\`  →  \`${newMember.displayName || "none"}\``,
-          inline: false,
-        });
-        if (fullOldDisplayName || oldMember.displayName === null)
-          return console.log("mem displayname null");
-        // await auditChannel.send({ embeds: [auditEmbed] });
-        await auditChannel2.send({ embeds: [auditEmbed] });
-      });
-    }
-
-    if (oldMember.user.username !== newMember.user.username) {
-      oldMember.fetch().then(async (fullOldMember) => {
-        const fullOldUsername = fullOldMember.user.username;
-        auditEmbed.addFields({
-          name: `Username Updated`,
-          value: `Name: \`${
-            fullOldUsername || oldMember.user.username || "none"
-          }\`  →  \`${newMember.user.username || "none"}\``,
-          inline: false,
-        });
-        if (fullOldUsername || oldMember.user.username === null)
-          return console.log("mem username null");
-        // await auditChannel.send({ embeds: [auditEmbed] });
-        await auditChannel2.send({ embeds: [auditEmbed] });
-      });
-    }
-
-    if (oldMember.avatar !== newMember.avatar) {
-      oldMember.fetch().then(async (fullOldMember) => {
-        const fullOldPfp = fullOldMember.avatarURL({ size: 512 });
-        const oldPfp = oldMember.avatarURL({ size: 512 });
-        const pfp = newMember.avatarURL({ size: 64 });
-        auditEmbed.setImage(`${pfp}`).addFields({
-          name: `Profile picture updated`,
-          value: `[Old Pfp](<${
-            fullOldPfp || oldPfp
-          }>)  →  [New Pfp](<${newMember.avatarURL({ size: 512 })}>)`,
-          inline: false,
-        });
-        if (newMember.avatar === null) return console.log("mem avatar = null");
-        if (fullOldPfp && oldPfp === null)
-          return console.log("mem avatar old null");
-        // await auditChannel.send({ embeds: [auditEmbed] });
-        await auditChannel2.send({ embeds: [auditEmbed] });
-      });
-    }
-
-    if (oldMember.user.avatar !== newMember.user.avatar) {
-      oldMember.fetch().then(async (fullOldMember) => {
-        const fullOldUserPfp = fullOldMember.user.avatarURL({ size: 512 });
-        const oldUserPfp = oldMember.user.avatarURL({ size: 512 });
-        const UserPfp = newMember.avatarURL({ size: 64 });
-        auditEmbed.setImage(`${UserPfp}`).addFields({
-          name: `Profile picture updated`,
-          value: `[Old Pfp](<${
-            fullOldUserPfp || oldUserPfp
-          }>)  →  [New Pfp](<${newMember.user.avatarURL({ size: 512 })}>)`,
-          inline: false,
-        });
-        if (newMember.user.avatar === null)
-          return console.log("memusr avatar = null");
-        if (fullOldUserPfp && oldUserPfp === null)
-          return console.log("memusr avatar old null");
-        // await auditChannel.send({ embeds: [auditEmbed] });
-        await auditChannel2.send({ embeds: [auditEmbed] });
-      });
-    }
-
-    if (oldMember.user.banner !== newMember.user.banner) {
-      oldMember.fetch().then(async (fullOldMember) => {
-        const fullOldBanner = fullOldMember.user.bannerURL({ size: 512 });
-        const oldBanner = oldMember.user.bannerURL({ size: 512 });
-        const Banner = newMember.user.bannerURL({ size: 64 });
-        auditEmbed.setImage(`${Banner}`).addFields({
-          name: `Banner updated`,
-          value: `[Old Banner](<${
-            fullOldBanner || oldBanner
-          }>)  →  [New Banner](<${newMember.user.bannerURL({ size: 512 })}>)`,
-          inline: false,
-        });
-        if (newMember.user.banner === null)
-          return console.log("memusr banner = null");
-        if (fullOldBanner && oldBanner === null)
-          return console.log("memusr banner old null");
-        // await auditChannel.send({ embeds: [auditEmbed] });
-        await auditChannel2.send({ embeds: [auditEmbed] });
-      });
-    }
-  });
-  client.on(Events.UserUpdate, async (oldUser, newUser) => {
-    const logGuild = await client.guilds.fetch(process.env.GUILD_ID);
-    const data = await Audit_Log.findOne({ Guild: logGuild }); // can't fetch guild this way?
-    let logID;
-    if (data) {
-      logID = data.Channel;
-    } else {
-      return;
-    }
-    if (oldUser.bot || newUser.bot) return;
-    if (oldUser.system || newUser.system) return;
-    if (oldUser.partial) console.log("oldUser = partial");
-
-    // const data = process.env.GUILD_ID; // gettin guild this way means it only works on FKZ
-
-    const auditEmbed = new EmbedBuilder()
-      .setColor("#ff00b3")
-      .setTimestamp()
-      .setFooter({ text: "FKZ Log System" })
-      .setTitle("User Updated")
-      .addFields({
-        name: "User:",
-        value: `<@${newUser.id}> - ${newUser.username}`,
-        inline: false,
-      });
-
-    const logschatID = await client.channels.fetch(process.env.LOGS_CHAT_ID);
-
-    const auditChannel = client.channels.cache.get(logID);
-    if (auditChannel === undefined) console.log("usr ndefined?", auditChannel);
-    const auditChannel2 = client.channels.cache.get(logschatID);
-    if (auditChannel === undefined)
-      console.log("usr undefined? 2", auditChannel2);
-
-    if (oldUser.username !== newUser.username) {
-      oldUser.fetch().then(async (fullOldUser) => {
-        const fullOldUsername = fullOldUser.username;
-        auditEmbed.addFields({
-          name: `Username Updated`,
-          value: `Name: \`${
-            fullOldUsername || oldUser.username || "none"
-          }\`  →  \`${newUser.username || "none"}\``,
-          inline: false,
-        });
-        if (newUser.username === null) return console.log("usr username null");
-        if (fullOldUsername || oldUser.username === null)
-          return console.log("usr username old null");
-        // await auditChannel.send({ embeds: [auditEmbed] });
-        await auditChannel2.send({ embeds: [auditEmbed] });
-      });
-    }
-
-    if (oldUser.avatar !== newUser.avatar) {
-      oldUser.fetch().then(async (fullOldUser) => {
-        const fullOldPfp = fullOldUser.avatarURL({ size: 512 });
-        const oldPfp = oldUser.avatarURL({ size: 512 });
-        const UserPfp = newUser.avatarURL({ size: 64 });
-        auditEmbed.setImage(`${UserPfp}`).addFields({
-          name: `Profile picture updated`,
-          value: `[Old Pfp](<${
-            fullOldPfp || oldPfp
-          }>)  →  [New Pfp](<${newUser.avatarURL({ size: 512 })}>)`,
-          inline: false,
-        });
-        if (newUser.avatar === null) return console.log("usr avatar = null");
-        if (newUser.avatar === undefined)
-          return console.log("avatar = undefined");
-        if (fullOldPfp && oldPfp === null)
-          return console.log("usr avatar old = null");
-        // await auditChannel.send({ embeds: [auditEmbed] });
-        await auditChannel2.send({ embeds: [auditEmbed] });
-      });
-    }
-
-    if (oldUser.banner !== newUser.banner) {
-      oldUser.fetch().then(async (fullOldUser) => {
-        const fullOldBanner = fullOldUser.bannerURL({ size: 512 });
-        const oldBanner = oldUser.bannerURL({ size: 512 });
-        const Banner = newUser.bannerURL({ size: 64 });
-        auditEmbed.setImage(`${Banner}`).addFields({
-          name: `Banner updated`,
-          value: `[Old Banner](<${
-            fullOldBanner || oldBanner
-          }>)  →  [New Banner](<${newUser.bannerURL({ size: 512 })}>)`,
-          inline: false,
-        });
-        if (newUser.banner === null) return console.log("banner = null");
-        if (newUser.banner === undefined)
-          return console.log("banner = undefined");
-        if (fullOldBanner && oldBanner === null)
-          return console.log("usr banner old = null");
-        // await auditChannel.send({ embeds: [auditEmbed] });
-        await auditChannel2.send({ embeds: [auditEmbed] });
-      });
-    }
-  });
-});
-client.on(Events.GuildMemberAdd, async (member) => {
-  const data = await Audit_Log.findOne({
-    Guild: member.guild.id,
-  });
-
-  let logID;
-  if (data) {
-    logID = data.Channel;
-  } else {
-    return;
-  }
-
-  const newInvites = await member.guild.invites.fetch();
-  const oldInvites = invites.get(member.guild.id);
-
-  const invite = newInvites.find((i) => i.uses > oldInvites.get(i.code));
-
-  const auditChannel = client.channels.cache.get(logID);
-  const auditEmbed = new EmbedBuilder()
-    .setColor("#ff00b3")
-    .setTimestamp()
-    .setTitle(`${member.user.username} Has Joined the Server!`)
-    .setAuthor({ name: `Member Joined` })
-    .setImage("https://femboy.kz/images/wide.png")
-    .setFooter({ text: "FKZ Log System" });
-
-  if (!invite) {
-    auditEmbed.setDescription(
-      `<@${member.user.id}> joined the server using an \`unknown invite\`. This could mean they used a vanity invite link if the server has one.`
-    );
-    return await auditChannel.send({ embeds: [auditEmbed] });
-  }
-
-  const inviter = await client.users.fetch(invite.inviter.id);
-
-  const auditEmbed2 = new EmbedBuilder()
-    .setColor("#ff00b3")
-    .setTimestamp()
-    .setTitle(`${member.user.username} Has Joined the Server!`)
-    .setDescription(
-      `<@${member.user.id}> Joined the server using the invite: \`${invite.code}\` Which was created by: ${inviter.tag}.\nThe invite has been used \`${invite.uses}\` times since it was created.`
-    )
-    .setAuthor({ name: `Member Joined` })
-    .setImage("https://femboy.kz/images/wide.png")
-    .setFooter({ text: "FKZ Log System" });
-
-  const auditEmbed3 = new EmbedBuilder()
-    .setColor("#ff00b3")
-    .setTimestamp()
-    .setTitle(`${member.user.username} Has Joined the Server!`)
-    .setDescription(
-      `<@${member.user.id}> Joined the server, but the invite used cannot be found. `
-    )
-    .setAuthor({ name: `Member Joined` })
-    .setImage("https://femboy.kz/images/wide.png")
-    .setFooter({ text: "FKZ Log System" });
-
-  inviter
-    ? auditChannel.send({ embeds: [auditEmbed2] })
-    : auditChannel.send({ embeds: [auditEmbed3] });
-});
-client.on(Events.GuildMemberRemove, async (member) => {
-  const data = await Audit_Log.findOne({
-    Guild: member.guild.id,
-  });
-  let logID;
-  if (data) {
-    logID = data.Channel;
-  } else {
-    return;
-  }
-  const auditChannel = client.channels.cache.get(logID);
-
-  const auditEmbed = new EmbedBuilder()
-    .setColor("#ff00b3")
-    .setTimestamp()
-    .setAuthor({ name: `Member Left` })
-    .setFooter({ text: "FKZ Log System" })
-    .setTitle(`${member.user.username} has left the server`)
-    .setDescription(`${member} has left the Server`);
-  await auditChannel.send({ embeds: [auditEmbed] });
-});
-client.on(Events.GuildMemberUpdate, async (oldMember, newMember) => {
-  const data = await Audit_Log.findOne({
-    Guild: oldMember.guild.id,
-  });
-  let logID;
-  if (data) {
-    logID = data.Channel;
-  } else {
-    return;
-  }
-
-  if (oldMember.user.bot || newMember.user.bot) return;
-  if (oldMember.user.system || newMember.user.system) return;
-  if (oldMember.partial) {
-    await oldMember.fetch().catch(() => {});
-  }
-
-  try {
-    const auditEmbed = new EmbedBuilder()
-      .setColor("#ff00b3")
-      .setTimestamp()
-      .setFooter({ text: "FKZ Log System" })
-      .setTitle("Member Updated");
-
-    const auditChannel = client.channels.cache.get(logID);
-
-    if (oldMember.roles.cache.size > newMember.roles.cache.size) {
-      oldMember.roles.cache.forEach((role) => {
-        if (!newMember.roles.cache.has(role.id)) {
-          auditEmbed.addFields(
-            {
-              name: "Role Removed: ",
-              value: `${role}`,
-              inline: false,
-            },
-            {
-              name: "User:",
-              value: `<@${newMember.user.id}> - ${newMember.user.username}`,
-              inline: false,
-            }
-          );
-          auditChannel.send({ embeds: [auditEmbed] });
-        }
-      });
-    }
-    if (oldMember.roles.cache.size < newMember.roles.cache.size) {
-      newMember.roles.cache.forEach((role) => {
-        if (!oldMember.roles.cache.has(role.id)) {
-          auditEmbed.addFields(
-            {
-              name: "Role Added: ",
-              value: `${role}`,
-              inline: false,
-            },
-            {
-              name: "User:",
-              value: `<@${newMember.user.id}> - ${newMember.user.username}`,
-              inline: false,
-            }
-          );
-          auditChannel.send({ embeds: [auditEmbed] });
-        }
-      });
-    }
-  } catch (err) {
-    console.log("my balls", err);
-  }
-});
-
-// ------------------------------------------------------------------------ AUTOROLE
+// AUTO ROLES //
 const autorole = require("./Schemas/autorole");
 
 client.on(Events.GuildMemberAdd, async (member) => {
